Fixed some issues

+++ src/controllers/dislike.controller.js
@@ -1,11 +1,11 @@
 import mongoose, { isValidObjectId } from "mongoose";
-import { Like } from "../models/like.model.js";
 import { Dislike } from "../models/dislike.model.js";
+import { Like } from "../models/like.model.js";
 import { ApiError } from "../utils/ApiError.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 
-const toggleVideoLike = asyncHandler(async (req, res) => {
+const toggleVideoDislike = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
 
     if (!videoId) {
@@ -16,35 +16,35 @@ const toggleVideoLike = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Invalid video id");
     }
 
-    let like = await Like.findOneAndDelete({
-        likedBy: req.user._id,
+    let dislike = await Dislike.findOneAndDelete({
+        dislikedBy: req.user._id,
         video: videoId,
     });
 
-    if (!like) {
-        await Dislike.findOneAndDelete({
-            dislikedBy: req.user._id,
+    if (!dislike) {
+        await Like.findOneAndDelete({
+            likedBy: req.user._id,
             video: videoId,
         });
-        like = await Like.create({
-            likedBy: req.user._id,
+        dislike = await Dislike.create({
+            dislikedBy: req.user._id,
             video: videoId,
         });
     }
 
-    if (!like) {
+    if (!dislike) {
         throw new ApiError(
             500,
-            "Something went wrong while toggling video like"
+            "Something went wrong while toggling video dislike"
         );
     }
 
     res.status(200).json(
-        new ApiResponse(200, like, "Video like toggled successfully")
+        new ApiResponse(200, dislike, "Video dislike toggled successfully")
     );
 });
 
-const toggleCommentLike = asyncHandler(async (req, res) => {
+const toggleCommentDislike = asyncHandler(async (req, res) => {
     const { commentId } = req.params;
 
     if (!commentId) {
@@ -55,35 +55,35 @@ const toggleCommentLike = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Invalid comment id");
     }
 
-    let like = await Like.findOneAndDelete({
+    let dislike = await Dislike.findOneAndDelete({
         comment: commentId,
-        likedBy: req.user?._id,
+        dislikedBy: req.user?._id,
     });
 
-    if (!like) {
-        await Dislike.findOneAndDelete({
+    if (!dislike) {
+        await Like.findOneAndDelete({
             comment: commentId,
-            dislikedBy: req.user?._id,
-        });
-        like = await Like.create({
             likedBy: req.user?._id,
+        });
+        dislike = await Dislike.create({
+            dislikedBy: req.user?._id,
             comment: commentId,
         });
     }
 
-    if (!like) {
+    if (!dislike) {
         throw new ApiError(
             500,
-            "Something went wrong while toggling comment like"
+            "Something went wrong while toggling comment dislike"
         );
     }
 
     res.status(200).json(
-        new ApiResponse(200, like, "Comment like toggled successfully")
+        new ApiResponse(200, dislike, "Comment dislike toggled successfully")
     );
 });
 
-const toggleTweetLike = asyncHandler(async (req, res) => {
+const toggleTweetDislike = asyncHandler(async (req, res) => {
     const { tweetId } = req.params;
 
     if (!tweetId) {
@@ -94,35 +94,35 @@ const toggleTweetLike = asyncHandler(async (req, res) => {
         throw new ApiError(400, "Invalid tweet id");
     }
 
-    let like = await Like.findOneAndDelete({
+    let dislike = await Dislike.findOneAndDelete({
         tweet: tweetId,
-        likedBy: req.user?._id,
+        dislikedBy: req.user?._id,
     });
 
-    if (!like) {
-        await Dislike.findOneAndDelete({
+    if (!dislike) {
+        await Like.findOneAndDelete({
             tweet: tweetId,
-            dislikedBy: req.user?._id,
-        });
-        like = await Like.create({
             likedBy: req.user?._id,
+        });
+        dislike = await Dislike.create({
+            dislikedBy: req.user?._id,
             tweet: tweetId,
         });
     }
 
-    if (!like) {
+    if (!dislike) {
         throw new ApiError(
             500,
-            "Something went wrong while toggling tweet like"
+            "Something went wrong while toggling tweet dislike"
         );
     }
 
     res.status(200).json(
-        new ApiResponse(200, like, "Tweet like toggled successfully")
+        new ApiResponse(200, dislike, "Tweet dislike toggled successfully")
     );
 });
 
-const getLikedVideos = asyncHandler(async (req, res) => {
+const getDislikedVideos = asyncHandler(async (req, res) => {
     if (!req.user?._id) {
         throw new ApiError(401, "Unauthorized");
     }
@@ -131,10 +131,10 @@ const getLikedVideos = asyncHandler(async (req, res) => {
         throw new ApiError(401, "Invalid user id");
     }
 
-    const likedVideos = await Like.aggregate([
+    const dislikedVideos = await Dislike.aggregate([
         {
             $match: {
-                likedBy: new mongoose.Types.ObjectId(req.user._id),
+                dislikedBy: new mongoose.Types.ObjectId(req.user._id),
                 video: { $ne: null },
             },
         },
@@ -201,16 +201,25 @@ const getLikedVideos = asyncHandler(async (req, res) => {
         },
     ]);
 
-    if (!likedVideos) {
+    if (!dislikedVideos) {
         throw new ApiError(
             500,
-            "Something went wrong while fetching liked videos"
+            "Something went wrong while fetching disliked videos"
         );
     }
 
     res.status(200).json(
-        new ApiResponse(200, likedVideos, "Liked videos fetched successfully")
+        new ApiResponse(
+            200,
+            dislikedVideos,
+            "Disliked videos fetched successfully"
+        )
     );
 });
 
-export { toggleCommentLike, toggleTweetLike, toggleVideoLike, getLikedVideos };
+export {
+    toggleCommentDislike,
+    toggleTweetDislike,
+    toggleVideoDislike,
+    getDislikedVideos,
+};
